% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{bonsaisitter-package}
\alias{bonsaisitter}
\alias{bonsaisitter-package}
\title{\packageTitle{bonsaisitter}}
\description{\packageDescription{bonsaisitter}}
\section{Package Content}{\packageIndices{bonsaisitter}}
\author{\packageAuthor{bonsaisitter}}
\section{Maintainer}{\packageMaintainer{bonsaisitter}}
\keyword{internal}
