% tinyrox says don't edit this manually, but it can't stop you!
\name{languages}
\alias{languages}
\alias{language_python}
\alias{language_cpp}
\title{Bundled tree-sitter grammars}
\usage{
languages()

language_python()

language_cpp()
}
\value{
A `tree_sitter_language`.
}
\description{
`language_r()`, `language_python()`, and `language_cpp()` return the bundled
grammar as a `tree_sitter_language` object, for use with [parser()].
}
