% tinyrox says don't edit this manually, but it can't stop you!
\name{node-child-count}
\alias{node_child_count}
\alias{node_named_child_count}
\title{Child counts}
\usage{
node_child_count(x)

node_named_child_count(x)
}
\arguments{
\item{x}{A `tree_sitter_node`.}
}
\value{
A single double.
}
\description{
Child counts
}
