% tinyrox says don't edit this manually, but it can't stop you!
\name{node_show_s_expression}
\alias{node_show_s_expression}
\title{Show a node's s-expression}
\usage{
node_show_s_expression(x)
}
\arguments{
\item{x}{A `tree_sitter_node`.}
}
\value{
`x`, invisibly, called for its side effect of printing.
}
\description{
Show a node's s-expression
}
