% tinyrox says don't edit this manually, but it can't stop you!
\name{parser_parse}
\alias{parser_parse}
\title{Parse text into a syntax tree}
\usage{
parser_parse(x, text, ...)
}
\arguments{
\item{x}{A `tree_sitter_parser`.}

\item{text}{A single string to parse.}

\item{...}{Unused.}
}
\value{
A `tree_sitter_tree`.
}
\description{
Parse text into a syntax tree
}
