% tinyrox says don't edit this manually, but it can't stop you!
\name{parser_set_language}
\alias{parser_set_language}
\title{Set a parser's language}
\usage{
parser_set_language(x, language)
}
\arguments{
\item{x}{A `tree_sitter_parser`.}

\item{language}{A `tree_sitter_language`.}
}
\value{
A new `tree_sitter_parser`.
}
\description{
Set a parser's language
}
