% tinyrox says don't edit this manually, but it can't stop you!
\name{query-counts}
\alias{query_pattern_count}
\alias{query_capture_count}
\alias{query_string_count}
\title{Query counts}
\usage{
query_pattern_count(x)

query_capture_count(x)

query_string_count(x)
}
\arguments{
\item{x}{A \code{tree_sitter_query}.}
}
\value{
A single double.
}
\description{
Query counts
}
\examples{
if (requireNamespace("treesitter.r", quietly = TRUE)) {
  q <- query(treesitter.r::language(), "(call function: (identifier) @fn)")
  c(query_pattern_count(q), query_capture_count(q), query_string_count(q))
}
}
