% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{treesitR-package}
\alias{treesitR}
\alias{treesitR-package}
\title{\packageTitle{treesitR}}
\description{\packageDescription{treesitR}}
\section{Package Content}{\packageIndices{treesitR}}
\author{\packageAuthor{treesitR}}
\section{Maintainer}{\packageMaintainer{treesitR}}
\keyword{internal}
