% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_cursor_depth}
\alias{ts_cursor_depth}
\title{Get cursor depth}
\usage{
ts_cursor_depth(cursor)
}
\arguments{
\item{cursor}{A \code{ts_cursor} object.}
}
\value{
Integer depth relative to the cursor's root node.
}
\description{
Get cursor depth
}
