% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_cursor_field_name}
\alias{ts_cursor_field_name}
\title{Get field name at cursor position}
\usage{
ts_cursor_field_name(cursor)
}
\arguments{
\item{cursor}{A \code{ts_cursor} object.}
}
\value{
A character string or \code{NA} if no field.
}
\description{
Get field name at cursor position
}
