% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_cursor_goto_first_child}
\alias{ts_cursor_goto_first_child}
\title{Move cursor to first child}
\usage{
ts_cursor_goto_first_child(cursor)
}
\arguments{
\item{cursor}{A \code{ts_cursor} object.}
}
\value{
\code{TRUE} if moved, \code{FALSE} if no children.
}
\description{
Move cursor to first child
}
