% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_cursor_goto_next_sibling}
\alias{ts_cursor_goto_next_sibling}
\title{Move cursor to next sibling}
\usage{
ts_cursor_goto_next_sibling(cursor)
}
\arguments{
\item{cursor}{A \code{ts_cursor} object.}
}
\value{
\code{TRUE} if moved, \code{FALSE} if no next sibling.
}
\description{
Move cursor to next sibling
}
