% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_cursor_goto_parent}
\alias{ts_cursor_goto_parent}
\title{Move cursor to parent}
\usage{
ts_cursor_goto_parent(cursor)
}
\arguments{
\item{cursor}{A \code{ts_cursor} object.}
}
\value{
\code{TRUE} if moved, \code{FALSE} if at root.
}
\description{
Move cursor to parent
}
