% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_cursor_new}
\alias{ts_cursor_new}
\title{Create a new tree cursor}
\usage{
ts_cursor_new(node)
}
\arguments{
\item{node}{A \code{ts_node} to start from.}
}
\value{
A \code{ts_cursor} object (external pointer).
}
\description{
A tree cursor allows efficient traversal of a syntax tree.
}
