% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_cursor_node}
\alias{ts_cursor_node}
\title{Get the current node of a cursor}
\usage{
ts_cursor_node(cursor)
}
\arguments{
\item{cursor}{A \code{ts_cursor} object.}
}
\value{
A \code{ts_node} object.
}
\description{
Get the current node of a cursor
}
