% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_language_python}
\alias{ts_language_python}
\title{Return the tree-sitter Python language grammar}
\usage{
ts_language_python()
}
\value{
An external pointer of class \code{"ts_language"}.
}
\description{
Returns an external pointer to the bundled Python language grammar
for use with \code{\link{ts_parser_set_language}}.
}
