% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_language_r}
\alias{ts_language_r}
\title{Return the tree-sitter R language grammar}
\usage{
ts_language_r()
}
\value{
An external pointer of class \code{"ts_language"}.
}
\description{
Returns an external pointer to the bundled R language grammar
for use with \code{\link{ts_parser_set_language}}.
}
