% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_child_by_field}
\alias{ts_node_child_by_field}
\title{Get a child node by field name}
\usage{
ts_node_child_by_field(node, name)
}
\arguments{
\item{node}{A \code{ts_node} object.}

\item{name}{Field name as a character string.}
}
\value{
A \code{ts_node} or \code{NULL}.
}
\description{
Get a child node by field name
}
