% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_child_count}
\alias{ts_node_child_count}
\title{Get the number of children of a node}
\usage{
ts_node_child_count(node)
}
\arguments{
\item{node}{A \code{ts_node} object.}
}
\value{
Integer.
}
\description{
Get the number of children of a node
}
