% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_children}
\alias{ts_node_children}
\title{Get all children of a node}
\usage{
ts_node_children(node, named = FALSE)
}
\arguments{
\item{node}{A \code{ts_node} object.}

\item{named}{If \code{TRUE}, return only named children.}
}
\value{
A list of \code{ts_node} objects.
}
\description{
Get all children of a node
}
