% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_is_named}
\alias{ts_node_is_named}
\title{Check if a node is named}
\usage{
ts_node_is_named(node)
}
\arguments{
\item{node}{A \code{ts_node} object.}
}
\value{
Logical.
}
\description{
Named nodes correspond to named rules in the grammar.
Anonymous nodes correspond to string literals.
}
