% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_named_child}
\alias{ts_node_named_child}
\title{Get a named child of a node by index}
\usage{
ts_node_named_child(node, i)
}
\arguments{
\item{node}{A \code{ts_node} object.}

\item{i}{Zero-based index among named children only.}
}
\value{
A \code{ts_node} or \code{NULL}.
}
\description{
Get a named child of a node by index
}
