% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_parent}
\alias{ts_node_parent}
\title{Get the parent of a node}
\usage{
ts_node_parent(node)
}
\arguments{
\item{node}{A \code{ts_node} object.}
}
\value{
A \code{ts_node} or \code{NULL} if at the root.
}
\description{
Get the parent of a node
}
