% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_prev_named_sibling}
\alias{ts_node_prev_named_sibling}
\title{Get the previous named sibling of a node}
\usage{
ts_node_prev_named_sibling(node)
}
\arguments{
\item{node}{A \code{ts_node} object.}
}
\value{
A \code{ts_node} or \code{NULL}.
}
\description{
Get the previous named sibling of a node
}
