% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_sexpr}
\alias{ts_node_sexpr}
\title{Get the S-expression representation of a node}
\usage{
ts_node_sexpr(node)
}
\arguments{
\item{node}{A \code{ts_node} object.}
}
\value{
A character string.
}
\description{
Get the S-expression representation of a node
}
