% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_start_byte}
\alias{ts_node_start_byte}
\title{Get the start byte offset of a node}
\usage{
ts_node_start_byte(node)
}
\arguments{
\item{node}{A \code{ts_node} object.}
}
\value{
Integer.
}
\description{
Get the start byte offset of a node
}
