% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_node_start_point}
\alias{ts_node_start_point}
\title{Get the start position of a node}
\usage{
ts_node_start_point(node)
}
\arguments{
\item{node}{A \code{ts_node} object.}
}
\value{
Named integer vector with \code{row} and \code{column} (0-based).
}
\description{
Get the start position of a node
}
