% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_tree_root_node}
\alias{ts_tree_root_node}
\title{Get the root node of a syntax tree}
\usage{
ts_tree_root_node(tree)
}
\arguments{
\item{tree}{A \code{ts_tree} object from \code{\link{ts_parse}}.}
}
\value{
A \code{ts_node} object.
}
\description{
Get the root node of a syntax tree
}
