% tinyrox says don't edit this manually, but it can't stop you!
\name{ts_tree_text}
\alias{ts_tree_text}
\title{Get the source text of a syntax tree}
\usage{
ts_tree_text(tree)
}
\arguments{
\item{tree}{A \code{ts_tree} object.}
}
\value{
The source string that was parsed.
}
\description{
Get the source text of a syntax tree
}
