% tinyrox says don't edit this manually, but it can't stop you!
\name{node-bytes}
\alias{node_start_byte}
\alias{node_end_byte}
\title{A node's start / end byte}
\usage{
node_start_byte(x)

node_end_byte(x)
}
\arguments{
\item{x}{A `tree_sitter_node`.}
}
\value{
A single double (0-indexed).
}
\description{
A node's start / end byte
}
