% tinyrox says don't edit this manually, but it can't stop you!
\name{node-children}
\alias{node_children}
\alias{node_named_children}
\title{All children of a node}
\usage{
node_children(x)

node_named_children(x)
}
\arguments{
\item{x}{A `tree_sitter_node`.}
}
\value{
A list of `tree_sitter_node`s.
}
\description{
All children of a node
}
