% tinyrox says don't edit this manually, but it can't stop you!
\name{node-points}
\alias{node_start_point}
\alias{node_end_point}
\title{A node's start / end point}
\usage{
node_start_point(x)

node_end_point(x)
}
\arguments{
\item{x}{A `tree_sitter_node`.}
}
\value{
A `tree_sitter_point`.
}
\description{
A node's start / end point
}
