% tinyrox says don't edit this manually, but it can't stop you!
\name{node-symbols}
\alias{node-symbols}
\alias{node_grammar_symbol}
\alias{node_parse_state}
\alias{node_next_parse_state}
\title{Node symbols and parse states}
\usage{
node-symbols(x)

node_grammar_symbol(x)

node_parse_state(x)

node_next_parse_state(x)
}
\arguments{
\item{x}{A `tree_sitter_node`.}
}
\value{
A single double.
}
\description{
Node symbols and parse states
}
