% tinyrox says don't edit this manually, but it can't stop you!
\name{node_child_by_field_id}
\alias{node_child_by_field_id}
\title{A child by field id}
\usage{
node_child_by_field_id(x, id)
}
\arguments{
\item{x}{A `tree_sitter_node`.}

\item{id}{A field id (see [language_field_id_for_name()]).}
}
\value{
A `tree_sitter_node`, or `NULL`.
}
\description{
A child by field id
}
