% tinyrox says don't edit this manually, but it can't stop you!
\name{node_child_by_field_name}
\alias{node_child_by_field_name}
\title{A child by field name}
\usage{
node_child_by_field_name(x, name)
}
\arguments{
\item{x}{A `tree_sitter_node`.}

\item{name}{A field name.}
}
\value{
A `tree_sitter_node`, or `NULL`.
}
\description{
A child by field name
}
