% tinyrox says don't edit this manually, but it can't stop you!
\name{node_descendant_count}
\alias{node_descendant_count}
\title{The number of descendants of a node}
\usage{
node_descendant_count(x)
}
\arguments{
\item{x}{A `tree_sitter_node`.}
}
\value{
A single double, including the node itself.
}
\description{
The number of descendants of a node
}
