% tinyrox says don't edit this manually, but it can't stop you!
\name{node_raw_s_expression}
\alias{node_raw_s_expression}
\title{The raw s-expression of a node}
\usage{
node_raw_s_expression(x)
}
\arguments{
\item{x}{A `tree_sitter_node`.}
}
\value{
A single string.
}
\description{
The raw s-expression of a node
}
