% tinyrox says don't edit this manually, but it can't stop you!
\name{query-counts}
\alias{query-counts}
\alias{query_capture_count}
\alias{query_string_count}
\title{Query counts}
\usage{
query-counts(x)

query_capture_count(x)

query_string_count(x)
}
\arguments{
\item{x}{A `tree_sitter_query`.}
}
\value{
A single double.
}
\description{
Query counts
}
