% tinyrox says don't edit this manually, but it can't stop you!
\name{tree_included_ranges}
\alias{tree_included_ranges}
\title{The tree's included ranges}
\usage{
tree_included_ranges(x)
}
\arguments{
\item{x}{A `tree_sitter_tree`.}
}
\value{
A list of `tree_sitter_range` objects.
}
\description{
The tree's included ranges
}
